Align prompt with first line of commands
When a command is longer than the width of the command box, it gets wrapped, but then the prompt is aligned with the center of the multi-line command. Instead it should be top-aligned to it shows up to the left of the first line of the command, like it would in a terminal.